New version III 3.0.
Added DEFAULT value for enumeration type.

Technically, this is a compatible new feature. DEFAULT is used as the name of an enum value, and is always there, but not in enumerations of values(). If previously there already was a value for DEFAULT, this stays (and will now be the DEFAULT ... but nowhere in the library the DEFAULT concept is actually used).